Move UnexportMainSymbol task to nativeplatform
This task is now shared between C++ and Swift, the package should
reflect that.
