+++ subprojects/language-native/src/main/java/org/gradle/language/nativeplatform/tasks/UnexportMainSymbol.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.language.swift.tasks;
+package org.gradle.language.nativeplatform.tasks;
 
 import org.gradle.api.Action;
 import org.gradle.api.DefaultTask;
